fix 2/1/2021

+++ database/migrations/2020_10_05_062807_create_order_details_table.php
@@ -15,7 +15,7 @@ class CreateOrderDetailsTable extends Migration
     {
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_id');
+            $table->unsignedBigInteger('order_id');
             $table->string('product_code',12);
             $table->string('product_name');
             $table->string('quantity_product');
