+++ database/migrations/2020_10_21_070342_create_payrolls_table.php
@@ -15,7 +15,7 @@ class CreatePayrollsTable extends Migration
     {
         Schema::create('payrolls', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('staff_id');
+            $table->unsignedBigInteger('staff_id');
             $table->integer('basic_salary');
             $table->timestamps();
         });
