+++ database/migrations/2020_12_25_210625_create_comments_table.php
@@ -15,8 +15,8 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->increments('id');
-            $table->smallInteger('user_id');
-            $table->smallInteger('product_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('product_id');
             $table->string('user_name');
             $table->smallInteger('point');
             $table->text('comment');
